Extract main function logic
The main function of the main package that builds the krakend binary
contains too much logic, witch makes it a bit harder for users to build
krakend by themselves.

This commit extracts the orchestration done by the main function into
`krakend.Main`. This way, the main package is greatly simplified, making
it easier for clients to build the krakend binary.
